Rename trailing semicolon rule
