+++ docs/user-guide/example-config.md
@@ -37,6 +37,7 @@ You might want to learn a little about [the conventions](https://github.com/styl
     "declaration-block-semicolon-space-after": "always"|"never"|"always-single-line"|"never-single-line",
     "declaration-block-semicolon-space-before": "always"|"never"|"always-single-line"|"never-single-line",
     "declaration-block-single-line-max-declarations": int,
+    "declaration-block-trailing-semicolon": "always"|"never",
     "declaration-colon-newline-after": "always"|"always-multi-line",
     "declaration-colon-space-after": "always"|"never"|"always-single-line",
     "declaration-colon-space-before": "always"|"never",
@@ -91,7 +92,6 @@ You might want to learn a little about [the conventions](https://github.com/styl
     "rule-no-shorthand-property-overrides": true,
     "rule-non-nested-empty-line-before": "always"|"never",
     "rule-properties-order": "alphabetical"|[],
-    "rule-trailing-semicolon": "always"|"never",
     "selector-class-pattern": string,
     "selector-combinator-space-after": "always"|"never",
     "selector-combinator-space-before": "always"|"never",
