+++ src/rules/declaration-block-trailing-semicolon/README.md
@@ -1,0 +1,63 @@
+# declaration-block-trailing-semicolon
+
+Require or disallow a trailing semicolon within declaration blocks.
+
+```css
+a { background: orange; color: pink; }
+/**                                ↑
+ *                    This semicolon */
+```
+
+The trailing semicolon is the *last* semicolon in a declaration block and it is optional.
+
+## Options
+
+`string`: `"always"|"never"`
+
+### `"always"`
+
+There *must always* be a trailing semicolon.
+
+The following patterns are considered warnings:
+
+```css
+a { color: pink }
+```
+
+```css
+a { background: orange; color: pink }
+```
+
+The following patterns are *not* considered warnings:
+
+```css
+a { color: pink; }
+```
+
+```css
+a { background: orange; color: pink; }
+```
+
+### `"never"`
+
+There *must never* be a trailing semicolon.
+
+The following patterns are considered warnings:
+
+```css
+a { color: pink; }
+```
+
+```css
+a { background: orange; color: pink; }
+```
+
+The following patterns are *not* considered warnings:
+
+```css
+a { color: pink }
+```
+
+```css
+a { background: orange; color: pink }
+```
